pyproject.toml
[project]
name = "paper-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "llama-index>=0.13.6",
    "notebook>=7.4.5",
    "python-dotenv>=1.1.1",
]
